refactor: split enduser service to satisfy 1200-line structure gate
Move auth/session and backfill into dedicated files so service.go stays under the CI line limit.
